Component code formatting
