--- frontend/web/src/components/Pages/AuthPage/Login/Login.module.scss
@@ -1,1 +1,0 @@
-.Login {}
