+++ frontend/web/src/components/Pages/AuthPage/Registration/Registration.lazy.js
@@ -1,0 +1,11 @@
+import React, { lazy, Suspense } from 'react';
+
+const LazyRegistration = lazy(() => import('./Registration'));
+
+const Registration = props => (
+  <Suspense fallback={null}>
+    <LazyRegistration {...props} />
+  </Suspense>
+);
+
+export default Registration;
