+++ frontend/web/src/components/Pages/AuthPage/Registration/Registration.test.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import { render, screen } from '@testing-library/react';
+import '@testing-library/jest-dom';
+import Registration from './Registration';
+
+describe('<Registration />', () => {
+  test('it should mount', () => {
+    render(<Registration />);
+
+    const registration = screen.getByTestId('Registration');
+
+    expect(registration).toBeInTheDocument();
+  });
+});
